% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_check_api_key.R
\name{check_api_key}
\alias{check_api_key}
\title{Check API key exists under the alias 'nz_stat_api_key' in .Renvrion file,
and throw error if it doesn't.}
\usage{
check_api_key()
}
\value{
An error or a string containing the API key.
}
\description{
Check API key exists under the alias 'nz_stat_api_key' in .Renvrion file,
and throw error if it doesn't.
}
