% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_check_api_key.R
\name{get_api_key}
\alias{get_api_key}
\title{Get the users NZ Statistics API key}
\usage{
get_api_key()
}
\value{
A string containing the API key.
}
\description{
Get the users NZ Statistics API key
}
