% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_odata.R
\name{get_odata}
\alias{get_odata}
\title{Call the Statistics NZ open data API}
\usage{
get_odata(endpoint, query_option, entity = "Observations")
}
\arguments{
\item{endpoint}{A string scaler referencing the service endpoint to be queried.
In this case a "service endpoint" can be thought of as a database. Endpoint names
can be found in the \emph{endpoint} column of the dataframe, which is generated using
the \code{get_odata_catalogue()} function.}

\item{query_option}{A query string used to filter and aggregate results. Refer to
\href{https://github.com/StatisticsNZ/open-data-api/blob/main/Example-R-requests.md}{Github}
for examples.}

\item{entity}{A string scaler referencing which entity in the data model to be used.
There are two options: "Observations" or "Resources".
Observations returns actual data observations relevant to your query.
Resources returns metadata associated with your query.}
}
\value{
A dataframe.
}
\description{
Call the Statistics NZ open data API
}
