% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_odata_catalogue.R
\name{get_odata_catalogue}
\alias{get_odata_catalogue}
\title{Information of available data on the Statistics NZ open data catalogue}
\usage{
get_odata_catalogue(slim_df = FALSE)
}
\arguments{
\item{slim_df}{A boolean TRUE/FALSE argument. When set to \strong{TRUE} the resultant dataframe
only contains columns: \emph{title}, \emph{description} and \emph{endpoint} of the available endpoints.
Otherwise, all information/metadata is returned. The default argument is set to \strong{FALSE}}
}
\value{
A dataframe
}
\description{
Information of available data on the Statistics NZ open data catalogue
}
