% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_odata_entities.R
\name{get_odata_entities}
\alias{get_odata_entities}
\title{Get catalogues available for each endpoint}
\usage{
get_odata_entities()
}
\value{
A dataframe.
}
\description{
Get catalogues available for each endpoint
}
\examples{
(stat_entities <- get_odata_entities())


}
