% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNZSAPI.R
\docType{package}
\name{R.NZSAPI}
\alias{R.NZSAPI}
\title{R.NZSAPI: Stats NZ API convenience functions}
\section{About}{

This package provides convenience functions to communicate with the Statistics NZ API.
To do so, you first require an API key, which can be obtained by following instructions
on the \href{https://github.com/StatisticsNZ/open-data-api}{Statistics NZ github page}.
}

\section{Key System Variable (Alias)}{

Once a key is obtained, \strong{it must be} stored as a system variable named "nz_stat_api_key"
in your ~/.Renviron file.
}

